Philosophers | OKAY!

+++ Rank03/Philosophers/includes/philo.h
@@ -6,7 +6,7 @@
 /*   By: marcosv2 <marcosv2@student.42.rio>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/19 22:01:56 by marcosv2          #+#    #+#             */
-/*   Updated: 2024/01/22 01:07:54 by marcosv2         ###   ########.fr       */
+/*   Updated: 2024/01/22 03:22:44 by marcosv2         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,6 @@
 # include "ph_includes.h"
 # include "ph_macros.h"
 # include "ph_structs.h"
-#include <unistd.h>
 
 // ph_inits.c
 int			ph_inits(t_info *info, int ac, char **av);
@@ -40,7 +39,7 @@ int			ph_usleep(useconds_t time);
 void		ph_mss(char *s, t_philo *philo);
 
 // ft_utils.c
-int			ft_atoi(const char *nptr);
+long		ft_atoi(const char *nptr);
 void		ft_putstr(char *s);
 void		ft_putendl_fd(char *s, int fd);
 int			ft_strcmp(char *s1, char *s2);
